requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-countries==7.2.1
django-heroku==0.3.1
django-timezone-field==4.1.2
gunicorn==20.0.4
Pillow==8.2.0
psycopg2==2.9.1
pytz==2019.3
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.0.1
